Move do_myquota(), do_mytables(), do_find_user(), do_user_tables(), do_describe_tables(), do_find_tables(), do_find_tables_with_column(), do_show_index(), do_load_table(), do_append_table() from easyaccess to do_utils file
